feat: add a minimal Button version
Added Button minmal prop to render a version without background.

Button code moved to button.tsx file and index.tsx used as a export. The main reason
for these changes was the absence of storybook props when used forwardRef. The fix was
to use a named export in the Button component.